feat: isolate media player in internal pipeline with clocksync
Refactor the media player block to use a separate internal GStreamer
pipeline behind an appsrc/appsink bridge, following the same pattern
as the WHIP ingest block. This fixes three problems:

1. No playback pacing: add clocksync(sync=true) so buffers flow at
   the correct rate when feeding a video compositor or live sink.
2. EOS propagation: EOS is now caught in the internal pipeline and
   triggers next-file advancement instead of stopping the main pipeline.
3. Seek disruption: seek operates only on the internal pipeline;
   downstream sees continuous timestamps via a computed offset.

Architecture:
  Internal pipeline: uridecodebin/urisourcebin -> clocksync -> appsink
  Main pipeline:     appsrc -> queue -> identity (video_out/audio_out)
  Bridge:            appsink callback -> push_sample with ts offset

The timestamp offset (shared AtomicI64 between audio/video) is computed
once from the first buffer and reset on seek, file switch, and resume
to prevent accumulated drift.

Also split mediaplayer.rs (1427 lines) into sub-modules following the
vision_mixer pattern: mod.rs, state.rs, builder.rs, bridge.rs,
definition.rs.

Frontend changes:
- Enable interactive seek bar in compact node view (click/drag)
- Enable seek slider in property inspector panel (show_full)
- Wire show_full into the properties panel for media_player blocks
- Add new "sync" block property (default true)
